requirements.txt
pandas~=2.2.3
duckduckgo_search~=8.0.1
newspaper3k~=0.2.8
Flask~=3.1.0
torch~=2.7.0
transformers~=4.51.3
lxml-html-clean==0.4.2
flask-cors==5.0.1